test(core): fold KubbDriver.generators.test.ts into KubbDriver.test.ts
Move the generator-dispatch characterization tests into the colocated
KubbDriver.test.ts so the driver has a single test file matching its source.
No test coverage changes: the same 6 cases and their helpers move over.
